coldcall-api/pyproject.toml
[project]
name = "coldcall-api"
version = "0.1.0"
description = "Cold Calling Agent SaaS backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

